Ajustes projeto 2

+++ src/main/java/com/dminer/controllers/PostController.java
@@ -19,6 +19,7 @@ import java.util.stream.Collectors;
 import javax.annotation.PostConstruct;
 import javax.validation.Valid;
 
+
 import org.aspectj.lang.annotation.Before;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -36,6 +37,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -50,9 +52,11 @@ import com.dminer.dto.PostDTO;
 import com.dminer.dto.PostRequestDTO;
 import com.dminer.dto.ReactDTO;
 import com.dminer.dto.SurveyRequestDTO;
+import com.dminer.dto.Token;
 import com.dminer.dto.UserDTO;
 import com.dminer.dto.UserReductDTO;
 import com.dminer.entities.Comment;
+import com.dminer.entities.Favorites;
 import com.dminer.entities.FileInfo;
 import com.dminer.entities.ReactUser;
 import com.dminer.entities.Post;
@@ -60,6 +64,7 @@ import com.dminer.entities.React;
 import com.dminer.entities.User;
 import com.dminer.enums.PostType;
 import com.dminer.repository.CommentRepository;
+import com.dminer.repository.FavoritesRepository;
 import com.dminer.repository.GenericRepositoryPostgres;
 import com.dminer.repository.ReactRepository;
 import com.dminer.repository.ReactUserRepository;
@@ -116,6 +121,11 @@ public class PostController {
 	@Autowired
 	private GenericRepositoryPostgres genericRepositoryPostgres;
 	
+	@Autowired
+	private FavoritesRepository favoritesRepository;
+
+
+
 	private Gson gson = new Gson();
 	
 	private String token;
@@ -169,7 +179,7 @@ public class PostController {
 			post.setType(PostType.EXTERNAL);
 		}
 
-		response.setData(postToDto(post, null));
+		response.setData(postToDto(post, null, null));
 		
 		post = postService.persist(post);
 
@@ -324,10 +334,15 @@ public class PostController {
 
 
 	@GetMapping(value = "/{id}")
-	public ResponseEntity<Response<PostDTO>> get(@PathVariable("id") int id) {
+	public ResponseEntity<Response<PostDTO>> get(@RequestHeader("x-access-token") Token token, @PathVariable("id") int id) {
 		
 		Response<PostDTO> response = new Response<>();
 		log.info("Recuperando Post {}", id);
+
+		if (token.naoPreenchido()) { 
+            response.getErrors().add("Token precisa ser informado");    		
+    		return ResponseEntity.badRequest().body(response);
+        }
 
 		Optional<Post> post = postService.findById(id);
 		if (!post.isPresent()) {
@@ -336,7 +351,7 @@ public class PostController {
 		}
 
 		Optional<List<Comment>> comment = commentService.findByPost(post.get());
-		PostDTO dto = postToDto(post.get(), comment.get());
+		PostDTO dto = postToDto(post.get(), comment.get(), token.getToken());
 		dto.setReacts(getReacts(post.get()));
 
 		response.setData(dto);
@@ -370,41 +385,49 @@ public class PostController {
 
 
 
-	private PostDTO postToDto(Post post, List<Comment> comments) {
+	private PostDTO postToDto(Post post, List<Comment> comments, String token) {
 		PostDTO dto = new PostDTO();
-		// post.getReacts().forEach(like -> {
-		// 	dto.getReacts().add(like.getLogin());
-		// });
-
+		
 		dto.setType(post.getType().toString());
 		dto.setId(post.getId());
 		dto.setContent(post.getContent());
 		dto.setTitle(post.getTitle());
 		dto.setAnexo(post.getAnexo());
 
+		List<Favorites> favorites = favoritesRepository.findAllByPost(post);
+		favorites.forEach(f -> {
+			dto.getFavorites().add(f.getUser().getLogin());
+		});
 
-		UserReductDTO user = userService.buscarUsuarioApiReduct(post.getLogin());      	
-		dto.setUser(user);
-        
-		if (comments != null && !comments.isEmpty()) {
-			comments = comments.stream()
-			.sorted(Comparator.comparing(Comment::getTimestamp).reversed())
-			.collect(Collectors.toList());
-
-			comments.forEach(comment -> {
-				dto.getComments().add(commentConverter.entityToDTO(post.getId(), user, comment));
-			});			
+		if (token != null) {
+			UserReductDTO user = userService.buscarUsuarioApiReduct(post.getLogin(), token);      	
+			dto.setUser(user);
+			
+			if (comments != null && !comments.isEmpty()) {
+				comments = comments.stream()
+				.sorted(Comparator.comparing(Comment::getTimestamp).reversed())
+				.collect(Collectors.toList());
+	
+				comments.forEach(comment -> {
+					dto.getComments().add(commentConverter.entityToDTO(post.getId(), user, comment));
+				});			
+			}
 		}
 		return dto;
 	}
 	
 	
 	@GetMapping("/all/{login}")
-	public ResponseEntity<Response<List<PostDTO>>> getAllByUser(@PathVariable("login") String login) {
+	public ResponseEntity<Response<List<PostDTO>>> getAllByUser(@RequestHeader("x-access-token") Token token, @PathVariable("login") String login) {
 		
 		Response<List<PostDTO>> response = new Response<>();
 		response.setData(new ArrayList<PostDTO>());
 		log.info("Recuperando todos os Post");
+
+		if (token.naoPreenchido()) { 
+            response.getErrors().add("Token precisa ser informado");    		
+    		return ResponseEntity.badRequest().body(response);
+        }
 
 		List<Post> posts = postService.findAllByLogin(login);
 		if (posts == null || posts.isEmpty()) {
@@ -418,7 +441,7 @@ public class PostController {
 
 		for (Post post : posts) {
 			Optional<List<Comment>> comment = commentService.findByPost(post);
-			PostDTO dto = postToDto(post, comment.get());
+			PostDTO dto = postToDto(post, comment.get(), token.getToken());
 			dto.setReacts(getReacts(post));
 			response.getData().add(dto);
 		}
@@ -427,7 +450,7 @@ public class PostController {
 	
 	
 	@GetMapping()
-	public ResponseEntity<Response<List<PostDTO>>> getAll() {
+	public ResponseEntity<Response<List<PostDTO>>> getAll(@RequestHeader("x-access-token") Token token) {
 		
 		Response<List<PostDTO>> response = new Response<>();
 		response.setData(new ArrayList<PostDTO>());
@@ -439,13 +462,18 @@ public class PostController {
 			return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
 		}
 
+		if (token.naoPreenchido()) { 
+            response.getErrors().add("Token precisa ser informado");    		
+    		return ResponseEntity.badRequest().body(response);
+        }
+
 		posts = posts.stream()
 		.sorted(Comparator.comparing(Post::getCreateDate).reversed())
 		.collect(Collectors.toList());
 
 		for (Post post : posts) {
 			Optional<List<Comment>> comment = commentService.findByPost(post);
-			PostDTO dto = postToDto(post, comment.get());
+			PostDTO dto = postToDto(post, comment.get(), token.getToken());
 			dto.setReacts(getReacts(post));
 			response.getData().add(dto);
 		}
@@ -458,12 +486,17 @@ public class PostController {
 
 	@GetMapping(value = "/search/{id}")
     @Transactional(timeout = 50000)
-    public ResponseEntity<Response<PostDTO>> searchById(@PathVariable Integer id, @RequestParam(name = "date", required = false) String date, @RequestParam(name = "user", required = false) String user) {
+    public ResponseEntity<Response<PostDTO>> searchById(@RequestHeader("x-access-token") Token token, @PathVariable Integer id, @RequestParam(name = "date", required = false) String date, @RequestParam(name = "user", required = false) String user) {
         
         Response<PostDTO> response = new Response<>();
         if (id == null) {
             response.getErrors().add("Id precisa ser informado");
             return ResponseEntity.badRequest().body(response);
+        }
+
+		if (token.naoPreenchido()) { 
+            response.getErrors().add("Token precisa ser informado");    		
+    		return ResponseEntity.badRequest().body(response);
         }
 
 		Optional<Post> optPost = postService.findById(id);
@@ -479,7 +512,14 @@ public class PostController {
 		
 		Optional<User> optUser = null;
 		Integer userId = null;
-		UserRestModel userRestModel = userService.carregarUsuariosApi(TokenService.getToken());
+		UserRestModel userRestModel = null;
+	
+		userRestModel = userService.carregarUsuariosApi(token.getToken());
+
+		if (userRestModel == null) {
+			response.getErrors().add("Não foi possível carregar os usuários do endpoint");
+			return ResponseEntity.badRequest().body(response); 
+		}
 
 		optUser = userService.findByLoginApi(userSearch, userRestModel.getOutput().getResult().getUsuarios());
 		if (!optUser.isPresent()) {
@@ -492,7 +532,7 @@ public class PostController {
         
         List<Comment> comments = genericRepositoryPostgres.searchCommentsByPostIdAndDateAndUser(new Post(id), date, optUser);
 
-        PostDTO dto = postToDto(optPost.get(), comments);
+        PostDTO dto = postToDto(optPost.get(), comments, token.getToken());
         
 		dto.setReacts(getReacts(optPost.get()));
 
@@ -529,7 +569,7 @@ public class PostController {
 	///api/post/search/all?date=&user=
 	@GetMapping(value = "/search/all")
     @Transactional(timeout = 50000)
-    public ResponseEntity<Response<List<PostDTO>>> searchAll(@RequestParam(name = "date", required = false) String date, @RequestParam(name = "user", required = false) String user) {
+    public ResponseEntity<Response<List<PostDTO>>> searchAll(@RequestHeader("x-access-token") Token token, @RequestParam(name = "date", required = false) String date, @RequestParam(name = "user", required = false) String user) {
         
         Response<List<PostDTO>> response = new Response<>();
         
@@ -537,8 +577,19 @@ public class PostController {
 
 		UserRestModel userRestModel = null;
 
+		if (token.naoPreenchido()) { 
+            response.getErrors().add("Token precisa ser informado");    		
+    		return ResponseEntity.badRequest().body(response);
+        }
+
 		if (user != null && !user.isBlank()) {
-			userRestModel = userService.carregarUsuariosApi(TokenService.getToken());
+			userRestModel = userService.carregarUsuariosApi(token.getToken());
+			
+			if (userRestModel == null) {
+				response.getErrors().add("Não foi possível carregar os usuários do endpoint");
+				return ResponseEntity.badRequest().body(response); 
+			}
+
 			userOpt = userService.findByLoginApi(user, userRestModel.getOutput().getResult().getUsuarios());
 			if (!userOpt.isPresent()) {
 				response.getErrors().add("Nenhum usuário encontrado");
@@ -556,14 +607,7 @@ public class PostController {
 		List<PostDTO> postsDto = new ArrayList<>();
         
 		for (Post p : posts) {
-			// List<Comment> comms = genericRepositoryPostgres.searchCommentsByPostIdAndDateAndUser(p, date, userOpt);
-			
-			PostDTO dto = postToDto(p, null);
-			// comms.forEach(c -> {
-				// CommentDTO commDto = commentConverter.entityToDTO(c);
-				// dto.getComments().add(commDto);
-			// });
-			
+			PostDTO dto = postToDto(p, null, token.getToken());
 			dto.setReacts(getReacts(p));
 			postsDto.add(dto);
 		}
@@ -575,8 +619,8 @@ public class PostController {
 	
 
 	// /post/like/{id}/{login}
-	@PutMapping("/like/{id}/{login}/{react}")
-	public ResponseEntity<Response<PostDTO>> likes(@PathVariable("id") Integer idPost, @PathVariable("login") String login, @PathVariable("react") String react) {
+	@PutMapping("/like/{id}/{login}/{react}/{toggle}")
+	public ResponseEntity<Response<PostDTO>> likes(@RequestHeader("x-access-token") Token token, @PathVariable("id") Integer idPost, @PathVariable("login") String login, @PathVariable("react") String react, @PathVariable("toggle") Boolean toggle) {
 		Response<PostDTO> response = new Response<>();
 		if (idPost == null) {
             response.getErrors().add("Id precisa ser informado");
@@ -589,15 +633,18 @@ public class PostController {
             return ResponseEntity.badRequest().body(response);
         }
 		
+		if (token.naoPreenchido()) { 
+            response.getErrors().add("Token precisa ser informado");    		
+    		return ResponseEntity.badRequest().body(response);
+        }
+
 		Post post = optPost.get();
 
 		if (reactUserRepository.existsByLoginAndPost(login, post)) {
+		// if (!toggle) {
 			ReactUser reactUser = reactUserRepository.findByLoginAndPost(login, post);
 			reactUserRepository.deleteById(reactUser.getId());
 			return ResponseEntity.ok().build();
-			
-			// response.getErrors().add("Este usuário já está associado a este post");
-            // return ResponseEntity.badRequest().body(response);
 		}
 		
 		React reactObj = reactRepository.findByReact(react);
@@ -609,7 +656,7 @@ public class PostController {
 
 		//post.getReacts().add(like);
 		post = postService.persist(post);
-		PostDTO dto = postToDto(post, null);
+		PostDTO dto = postToDto(post, null, token.getToken());
 		dto.setReacts(getReacts(post));
 		response.setData(dto);
